fix: common AssetRuntime to AssetExtension
