+++ EMS/common-bundle/src/Twig/CommonExtension.php
@@ -23,28 +23,18 @@ class CommonExtension extends AbstractExtension
     public function getFunctions(): array
     {
         return [
-            new TwigFunction('ems_asset_path', [AssetRuntime::class, 'assetPath'], ['is_safe' => ['html']]),
-            new TwigFunction('ems_json_file', [AssetRuntime::class, 'jsonFromFile']),
-            new TwigFunction('ems_asset_get_content', [AssetRuntime::class, 'getContent']),
-            new TwigFunction('ems_asset_head', [AssetRuntime::class, 'head']),
-            new TwigFunction('ems_asset_heads', [AssetRuntime::class, 'heads']),
             new TwigFunction('ems_html', [TextRuntime::class, 'emsHtml'], ['is_safe' => ['all']]),
             new TwigFunction('ems_http', [HttpClientRuntime::class, 'request']),
             new TwigFunction('ems_nested_search', [SearchRuntime::class, 'nestedSearch']),
             new TwigFunction('ems_analyze', [SearchRuntime::class, 'analyze']),
-            new TwigFunction('ems_image_info', [AssetRuntime::class, 'imageInfo']),
             new TwigFunction('ems_version', [InfoRuntime::class, 'version']),
             new TwigFunction('ems_uuid', UuidGenerator::random(...)),
             new TwigFunction('ems_store_read', [StoreDataRuntime::class, 'read']),
             new TwigFunction('ems_store_save', [StoreDataRuntime::class, 'save']),
             new TwigFunction('ems_store_delete', [StoreDataRuntime::class, 'delete']),
             new TwigFunction('ems_template_exists', [TemplateRuntime::class, 'templateExists']),
-            new TwigFunction('ems_file_from_archive', [AssetRuntime::class, 'fileFromArchive']),
-            new TwigFunction('ems_files_in_archive', [AssetRuntime::class, 'getFilesInArchive']),
             new TwigFunction('ems_core', [CoreBridgeRuntime::class, 'build']),
             new TwigFunction('ems_flash', [RequestRuntime::class, 'flash']),
-            new TwigFunction('ems_file_reader_data', [AssetRuntime::class, 'fileReaderGetData']),
-            new TwigFunction('ems_file_reader_cells', [AssetRuntime::class, 'fileReaderReadCells']),
             new TwigFunction('ems_check_ip', [RequestRuntime::class, 'checkIp']),
             new TwigFunction('ems_clear_http_caches', [HttpCacheRuntime::class, 'clearCaches']),
         ];
@@ -70,13 +60,10 @@ class CommonExtension extends AbstractExtension
             new TwigFilter('ems_markdown', Encoder::markdownToHtml(...), ['is_safe' => ['html']]),
             new TwigFilter('ems_slug', [Encoder::class, 'slug']),
             new TwigFilter('ems_stringify', Converter::stringify(...)),
-            new TwigFilter('ems_temp_file', [AssetRuntime::class, 'temporaryFile']),
-            new TwigFilter('ems_asset_average_color', [AssetRuntime::class, 'assetAverageColor'], ['is_safe' => ['html']]),
             new TwigFilter('ems_replace_regex', [TextRuntime::class, 'replaceRegex'], ['is_safe' => ['html']]),
             new TwigFilter('ems_dom_crawler', [TextRuntime::class, 'domCrawler']),
             new TwigFilter('ems_base64_encode', Base64::encode(...)),
             new TwigFilter('ems_base64_decode', Base64::decode(...)),
-            new TwigFilter('ems_hash', [AssetRuntime::class, 'hash']),
             new TwigFilter('ems_preg_match', Encoder::pregMatch(...)),
             new TwigFilter('ems_color', fn ($color) => new Color($color)),
             new TwigFilter('ems_link', EMSLink::fromText(...)),
