+++ EMS/core-bundle/src/Command/ExportDocumentsCommand.php
@@ -10,7 +10,7 @@ use EMS\CommonBundle\Helper\EmsFields;
 use EMS\CommonBundle\Service\ElasticaService;
 use EMS\CommonBundle\Storage\Service\StorageInterface;
 use EMS\CommonBundle\Storage\StorageManager;
-use EMS\CommonBundle\Twig\AssetRuntime;
+use EMS\CommonBundle\Twig\AssetExtension;
 use EMS\CoreBundle\Commands;
 use EMS\CoreBundle\Service\ContentTypeService;
 use EMS\CoreBundle\Service\DataService;
@@ -64,7 +64,7 @@ class ExportDocumentsCommand extends AbstractCommand
         protected readonly DataService $dataService,
         protected readonly ContentTypeService $contentTypeService,
         protected readonly EnvironmentService $environmentService,
-        protected readonly AssetRuntime $runtime,
+        protected readonly AssetExtension $assetExtension,
         private readonly ElasticaService $elasticaService,
         private readonly StorageManager $storageManager,
         protected readonly string $instanceId
@@ -256,7 +256,7 @@ class ExportDocumentsCommand extends AbstractCommand
         }
 
         $hash = $this->storageManager->saveFile($outZipPath, StorageInterface::STORAGE_USAGE_CONFIG);
-        $url = ($this->baseUrl ?? '').$this->runtime->assetPath(
+        $url = ($this->baseUrl ?? '').$this->assetExtension->assetPath(
             [
                 EmsFields::CONTENT_FILE_HASH_FIELD => $hash,
                 EmsFields::CONTENT_FILE_NAME_FIELD => 'export.zip',
